backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
web3>=6.14.0
httpx>=0.26.0
python-dateutil>=2.8.2
# OpenSky API - install from GitHub if needed: pip install git+https://github.com/openskynetwork/opensky-api.git
# Or use REST API fallback (already implemented)
